+++ app/helpers/application_helper.rb
@@ -139,7 +139,7 @@ module ApplicationHelper
   end
 
   def request_placeholder(doc_id, holding_id)
-    "<div class=\"location-services\"><a target=\"_blank\" class=\"request btn btn-xs btn-primary\" title=\"Requesting is not currently functional\" data-toggle=\"tooltip\" href=\"/request\">Request</a></div>"
+    "<div class=\"location-services\"><a target=\"_blank\" class=\"request btn btn-xs btn-primary\" title=\"View Options to Request this Title\" data-toggle=\"tooltip\" href=\"/requests/#{doc_id}\">Request</a></div>"
   end
 
   def holding_block_search document
@@ -252,48 +252,6 @@ module ApplicationHelper
 
   def voyager_url bibid
     "http://catalog.princeton.edu/cgi-bin/Pwebrecon.cgi?BBID=#{bibid}"
-  end
-
-  def current_patron? netid
-    return false unless netid
-    begin 
-      patron_record = Faraday.get "#{ENV['bibdata_base']}/patron/#{netid}"
-    rescue Faraday::Error::ConnectionFailed => e
-      logger.info("Unable to connect to #{ENV['bibdata_base']}")
-      return false
-    end
-
-    if patron_record.status == 403
-      logger.info("403 Not Authorized to Connect to Patron Data Service at #{ENV['bibdata_base']}/patron/#{netid}")
-      return false 
-    end
-    if patron_record.status == 404
-      logger.info("404 Patron #{netid} cannot be found in the Patron Data Service.")
-      return false 
-    end
-    patron = JSON.parse(patron_record.body).with_indifferent_access
-    logger.info("#{patron.to_hash}")
-    patron
-  end
-
-  def voyager_myaccount? patron
-    begin
-      voyager_account = Faraday.get "#{ENV['voyager_api_base']}/vxws/MyAccountService?patronId=#{patron[:patron_id]}&patronHomeUbId=1@DB"
-    rescue Faraday::Error::ConnectionFailed => e
-      logger.info("Unable to Connect to #{ENV['voyager_api_base']}")
-      return false
-    end
-    if voyager_account.status == 403
-      logger.info("403 Not Authorized to Connect to Voyager My Account Service.")
-      return false
-    end
-    if voyager_account.status == 404
-      logger.info("404 Patron id #{patron[:patron_id]} cannot be found in the Voyager My Account Service.")
-      return false 
-    end
-    account = VoyagerAccount.new(voyager_account.body)
-    logger.info("#{account.source_doc}")
-    account
   end
 
   # Setup item status
