+++ app/views/_account.html.erb
@@ -7,20 +7,20 @@
       </button>
       <ul class="dropdown-menu" role="menu">
         <li>
-          <%= link_to 'My Account', page_path('account') %>
+          <%= link_to 'Your Account', main_app.page_path('account') %>
         </li>
         <li>
           <%= render partial: 'blacklight/nav/saved_searches' %>
         </li>
         <li>
-          <%= link_to t('blacklight.header_links.logout'), destroy_user_session_path %>
+          <%= link_to t('blacklight.header_links.logout'), main_app.destroy_user_session_path %>
         </li>
       </ul>
     </li>
     <% end %>
     <% else %>
     <li>
-      <%= link_to t('blacklight.header_links.login'), user_omniauth_authorize_path(:cas) %>
+      <%= link_to t('blacklight.header_links.login'), main_app.user_omniauth_authorize_path(:cas) %>
     </li>
     <% end %>
 <% end %>
