+++ app/views/account/_available_items.html.erb
@@ -1,7 +1,7 @@
 <h2 class="section-heading">Items Available for Pickup</h2>
 <% if (!@account.avail_items.nil?) %>
   <div class="table-responsive">
-    <table class="table table-striped table-bordered account-available">
+    <table class="table table-striped table-bordered account--available_items">
       <thead>
         <tr>
           <th>Item Title</th>
