+++ app/views/blacklight/nav/_saved_searches.html.erb
@@ -1,0 +1,1 @@
+<%= link_to t('blacklight.header_links.saved_searches'), main_app.saved_searches_path %>
