+++ app/views/layouts/application.html.erb
@@ -9,7 +9,30 @@
   <!--[if lte IE 8]><%= javascript_include_tag 'flot/excanvas.min' %><![endif]-->
 </head>
 <body>
-<%= yield %>
+<%= render :partial => 'shared/header_navbar' %>
 
-</body>
+  <%= render partial: 'shared/ajax_modal' %> <!-- only render if page is using modal -->
+
+  <main role="main" id="main-container">
+    <div class="container">
+
+      <div class="col-xs-12 col-lg-12 alert alert-warning beta">
+        <p><strong>This is Work-in-Progress</strong><br />
+          There are features missing. Expect things to break.
+        </p>
+      </div>
+
+      <%= content_tag :h1, application_name, class: 'sr-only application-heading' %>
+
+      <%= render :partial=>'/flash_msg', layout: 'shared/flash_messages' %>
+
+      <div class="row">
+        <%= yield %>
+      </div>
+
+    </div>
+  </main>
+
+  <%= render :partial => 'shared/footer' %>
+  </body>
 </html>
