+++ app/views/pages/myaccount.html.erb
@@ -1,5 +1,5 @@
 <div class="col-md-12">
-  <h1 class="page-heading">My Account</h1>
+  <h1 class="page-heading">Your Account</h1>
   <p>Review your library account.</p>
 
   <h2 class="section-heading">Personal Information</h2>
