+++ config/requests.yml.tmpl
@@ -7,6 +7,7 @@ defaults: &defaults
   stackmap_base: https://library.princeton.edu/utils/map
   voyager_api_base: http://catalog.princeton.edu:7014
   voyager_ub_id: 1@DB1
+  pulsearch_base: https://pulsearch.princeton.edu
   
 development:
   <<: *defaults
